refactor(utils.py,workflow.py): mv write_to_gcs_bucket to utils
