+++ src/cpg_flow/utils.py
@@ -16,6 +16,7 @@ from random import choices
 from typing import Union, cast
 
 import coloredlogs
+from google.cloud import storage
 
 import hail as hl
 from hailtop.batch import ResourceFile
@@ -386,3 +387,22 @@ def make_job_name(
     if part:
         name += f', {part}'
     return name
+
+
+def write_to_gcs_bucket(contents, path: Path):
+    client = storage.Client()
+
+    if not str(path).startswith('gs:/'):
+        raise ValueError(f'Path {path} must be a GCS path')
+
+    path = str(path).removeprefix('gs:/').removeprefix('/')
+    bucket_name, blob_name = path.split('/', 1)
+
+    bucket = client.bucket(bucket_name)
+    if not bucket.exists():
+        raise ValueError(f'Bucket {bucket_name} does not exist')
+
+    blob = bucket.blob(blob_name)
+    blob.upload_from_string(contents)
+
+    return bucket_name, blob_name
